feat: fallback to environment variables in SSE mode
Prior to this commit, the MCP server wasn't respecting the
`GRAFANA_URL` and `GRAFANA_API_KEY` environment variables when running in
SSE mode, and was expecting them to be passed in via headers
(which is admittedly non-standard).

This commit adds a fallback to the environment variables if the
headers aren't present.

Relates to #40.